pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "investments-mcp"
version = "0.1.0"
description = "MCP server for Obsidian investment vault — trade operations and EODHD market data"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "uvicorn>=0.29.0",
    "requests>=2.31",
    "yfinance>=0.2.40",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
investments-mcp = "investments_mcp.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["investments_mcp*"]
